Split banking agent domain modules
